[GoPlay] Merge AwsIdp from goplay_auth_aws into the extractor
